Audio System Done, Respawn/Checkpoint Started
Music for the 2D section is added in its final form, and no scripting is needed for it to work, just the standard audio source and listener set up. Respawn/Checkpoint system has started being added, testing is difficult because something is stopping the player from falling past the grid.