Respawn & Checkpoints as far as I can go
The version of the main level I have is not complete, so I have finished the checkpointing and respawning system as much as I can with the version of the level I have. Whats left should only take me about 15 minutes to implement. Everything up to the Moby dome is done.